+++ src/textileEngine.ts
@@ -3,14 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-//import * as crypto from 'crypto';
-//import * as path from 'path';
 import { TextileJS, Token, Options as TextileJSConfig } from '../libs/textile-js/textile';
+//import * as path from 'path';
 import * as vscode from 'vscode';
 import { TextileContributionProvider as TextileContributionProvider } from './textileExtensions';
 import { Slugifier } from './slugify';
 import { SkinnyTextDocument } from './tableOfContentsProvider';
-//import { TextileFileExtensions, Schemes, isOfScheme } from './util/links';
+//import { hash } from './util/hash';
+//import { isOfScheme, TextileFileExtensions, Schemes } from './util/links';
 
 const UNICODE_NEWLINE_REGEX = /\u2028|\u2029/g;
 
@@ -253,9 +253,7 @@ export class TextileEngine {
 
 			const src = token.attrGet('src');
 			if (src) {
-				const hash = crypto.createHash('sha256');
-				hash.update(src);
-				const imgHash = hash.digest('hex');
+				const imgHash = hash(src);
 				token.attrSet('id', `image-hash-${imgHash}`);
 			}
 
@@ -309,7 +307,7 @@ export class TextileEngine {
 					if (uri.path[0] === '/') {
 						const root = vscode.workspace.getWorkspaceFolder(this.currentDocument!);
 						if (root) {
-							const fileUri = vscode.Uri.file(path.join(root.uri.fsPath, uri.fsPath));
+							const fileUri = vscode.Uri.joinPath(root.uri, uri.fsPath);
 							uri = fileUri.with({
 								scheme: uri.scheme,
 								fragment: uri.fragment,