+++ preview-src/index.ts
@@ -24,6 +24,7 @@ const state = {
 	...(typeof originalState === 'object' ? originalState : {}),
 	...getData<any>('data-state')
 };
+
 // Make sure to sync VS Code state here
 vscode.setState(state);
 
@@ -193,12 +194,14 @@ document.addEventListener('click', event => {
 				return;
 			}
 
-			// Pass through known schemes
-			if (passThroughLinkSchemes.some(scheme => node.href.startsWith(scheme))) {
-				return;
+			let hrefText = node.getAttribute('data-href');
+			if (!hrefText) {
+				// Pass through known schemes
+				if (passThroughLinkSchemes.some(scheme => node.href.startsWith(scheme))) {
+					return;
+				}
+				hrefText = node.getAttribute('href');
 			}
-
-			const hrefText = node.getAttribute('data-href') || node.getAttribute('href');
 
 			// If original link doesn't look like a url, delegate back to VS Code to resolve
 			if (!/^[a-z\-]+:/i.test(hrefText)) {