+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "vscode-textile-preview" extension will be documented
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [NOT RELEASED]
+### Changed
+- Update to vscode-1.47.3, without browser.webpack nor sourcemap removing
+
 
 ## [1.3.4] - 20200801
 ### Fixed