+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update to vscode-1.47.3, without browser.webpack nor sourcemap removing
 - [PREVIEW] Activate experimental custom editor
+- Update to vscode 1.48.0
 
 
 ## [1.3.4] - 20200801